prompt-template-test-python/pyproject.toml
[project]
name = "prompt-template-test-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dashscope>=1.25.14",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.20",
    "langchain-openai>=1.1.11",
    "pymilvus>=2.6.10",
    "python-dotenv>=1.2.2",
]
